fix: תיקוני ריוויו — הקשחת DB, דיוק כספי, מניעת race condition
- תיעוד OpenAPI באנגלית ב-auto_block.py (סטנדרט אחיד)
- Numeric(10,2) במקום Float ל-auto_block_min_debt (דיוק כספי)
- CHECK constraints ב-DB: grace_months 1-12, min_debt >= 0
- אינדקס חלקי idx_stations_active_auto_block למשימה היומית
- IntegrityError fallback + savepoint בחסימה אוטומטית (race condition)
- ולידציה לפני mutation ב-update_auto_block_settings (מונע dirty state)
- הסרת stations_skipped שלא בשימוש

כל 1057 הבדיקות עוברות.

+++ app/db/models/station.py
@@ -5,7 +5,7 @@ Station Model - תחנת משלוחים
 לכל תחנה יש בעלים (station owner) וסדרנים (dispatchers) שמנהלים את המשלוחים.
 """
 from datetime import datetime
-from sqlalchemy import Column, Integer, BigInteger, String, DateTime, Float, Boolean, ForeignKey, JSON
+from sqlalchemy import Column, Integer, BigInteger, String, DateTime, Float, Boolean, ForeignKey, JSON, Numeric
 from sqlalchemy.dialects.postgresql import JSONB
 from sqlalchemy.orm import relationship
 
@@ -41,7 +41,7 @@ class Station(Base):
     # הגדרות חסימה אוטומטית [סעיף 10]
     auto_block_enabled = Column(Boolean, default=True, nullable=False, server_default="true")
     auto_block_grace_months = Column(Integer, default=2, nullable=False, server_default="2")  # מספר חודשים רצופים לפני חסימה
-    auto_block_min_debt = Column(Float, default=0.0, nullable=False, server_default="0")  # סף חוב מינימלי (₪)
+    auto_block_min_debt = Column(Numeric(10, 2), default=0.0, nullable=False, server_default="0")  # סף חוב מינימלי (₪)
 
     created_at = Column(DateTime, default=datetime.utcnow)
     updated_at = Column(DateTime, default=datetime.utcnow, onupdate=datetime.utcnow)
